backend/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
supabase>=2.5.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
openai>=1.30.0
fpdf2>=2.8.0
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.40.0
